Modifying last committed code
A rebuild should not be triggered when a rules file changes
because remodel is not intelligent enough to discern if a
change was harmless or not. A simple whitespace change in
rules file should not trigger a rebuild of whole dependency
tree.

Instead now a "-B" option has been provided similar to make
where a rebuild of the whole dependency tree can be forced.